Break out Transform, Model, and math to separate modules
